refactor(timetable): consolidate POST /substitute into the deviations path (#1886)
- endpoint removed; the Betreuungsplan gap-fill quick action now calls
  POST /instances/{id}/deviations (same day-wide semantics, one write path)
- classifySubstitute + shared helpers stay in api/timetable, writes live in
  services/schedule; route-table golden regenerated, complexity ratchet entry
  removed; endpoint tests removed with the endpoint, the active-instance
  supervisor-swap scenario is ported to the deviations endpoint
