+++ backend/api/timetable/substitute.go
@@ -1,56 +1,28 @@
-// Package timetable — WP-B12 substitute flow.
+// Package timetable — shared substitute/deviation classification helpers.
 //
-//	POST /api/timetable/substitute
-//
-// Body: {absent_staff_id, substitute_staff_id, date}. For every instance_staff
-// row of the absent staff on the given date, mark the original is_absent=true
-// and insert a substitute row. Idempotent on replay; atomic on 409.
-//
-// substitute_staff_id is optional (#1840): when omitted the request is
-// "absent-only" — the staff is marked absent and each affected position is left
-// open (markAbsentOnly). No substitute row is created. Everything below applies
-// to the full substitute (both ids present) path.
-//
-// Atomicity note: the TenantTxMiddleware rolls the request transaction back
-// only on 5xx (see tenant/http_middleware.go:40). A 409 rendered mid-handler
-// would therefore commit any writes made before the 409. We avoid that with a
-// strict two-phase flow:
-//
-//	Phase A (Dry-Run): classify every target instance without writing.
-//	                   A single 409 case aborts the whole request.
-//	Phase B (Apply):   only reached when Phase A validates cleanly. All writes
-//	                   happen here, so a 409 never commits partial state.
-//
-// Permission: SchedulesManage. Same tenant tx as other /instances routes.
+// The former POST /api/timetable/substitute endpoint was consolidated into
+// POST /instances/{id}/deviations (#1886): its only frontend caller (the
+// Betreuungsplan gap-fill quick action) now sends the same day-wide
+// substitution through the atomic deviations save. What remains here is the
+// pure classification logic (classifySubstitute), the shared response row
+// (AffectedInstance), the time-conflict advisory builder, and the SSE
+// broadcast helper — all consumed by deviations.go. The write path lives in
+// services/schedule (ApplyAbsence/ApplySubstitute, #1886).
 package timetable
 
 import (
 	"context"
-	"encoding/json"
-	"errors"
 	"fmt"
 	"log/slog"
-	"net/http"
 	"strings"
-	"time"
 	"unicode/utf8"
 
-	"github.com/moto-nrw/project-phoenix/api/common"
 	"github.com/moto-nrw/project-phoenix/internal/timezone"
-	"github.com/moto-nrw/project-phoenix/models/base"
 	scheduleModel "github.com/moto-nrw/project-phoenix/models/schedule"
 	"github.com/moto-nrw/project-phoenix/realtime"
 	scheduleSvc "github.com/moto-nrw/project-phoenix/services/schedule"
 	"github.com/moto-nrw/project-phoenix/tenant"
 )
-
-// substituteRequest is the POST body shape.
-type substituteRequest struct {
-	AbsentStaffID     int64   `json:"absent_staff_id"`
-	SubstituteStaffID int64   `json:"substitute_staff_id"`
-	Date              string  `json:"date"`
-	Reason            *string `json:"reason,omitempty"` // optional "why" for the absence (#1840)
-}
 
 // substituteActionType is the stable per-instance action string returned in
 // the response. Callers switch on it rather than parsing messages.
@@ -76,15 +48,6 @@ type AffectedInstance struct {
 	Action     string `json:"action"`
 }
 
-// SubstituteResponse is the 200 body.
-type SubstituteResponse struct {
-	AbsentStaffID     int64                                `json:"absent_staff_id"`
-	SubstituteStaffID int64                                `json:"substitute_staff_id"`
-	Date              string                               `json:"date"`
-	AffectedInstances []AffectedInstance                   `json:"affected_instances"`
-	Warnings          []scheduleSvc.SubstituteTimeConflict `json:"warnings"`
-}
-
 // plannedOp holds one classified target instance and the decision of what
 // Phase B should do with it. Both pointer fields are always populated —
 // Phase A pushes an op only after successfully loading the instance and
@@ -102,385 +65,6 @@ func (op plannedOp) writeOp() scheduleSvc.SubstituteWriteOp {
 		OrigRow:  op.origRow,
 		Action:   op.action,
 	}
-}
-
-// substitute handles POST /api/timetable/substitute.
-func (rs *Resource) substitute(w http.ResponseWriter, r *http.Request) {
-	ctx := r.Context()
-
-	// --- Parse + validate the request body --------------------------------
-	var req substituteRequest
-	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
-		common.RenderError(w, r, common.ErrorInvalidRequest(errors.New("invalid JSON body")))
-		return
-	}
-	if req.AbsentStaffID <= 0 {
-		common.RenderError(w, r, common.ErrorInvalidRequest(errors.New("absent_staff_id is required")))
-		return
-	}
-	// substitute_staff_id is optional (#1840): omitting it (0) marks the staff
-	// absent and leaves the position open. A negative id is malformed.
-	if req.SubstituteStaffID < 0 {
-		common.RenderError(w, r, common.ErrorInvalidRequest(errors.New("substitute_staff_id must be a positive id")))
-		return
-	}
-	if req.SubstituteStaffID > 0 && req.AbsentStaffID == req.SubstituteStaffID {
-		common.RenderError(w, r, common.ErrorInvalidRequest(errors.New("absent and substitute staff must differ")))
-		return
-	}
-	if req.Date == "" {
-		common.RenderError(w, r, common.ErrorInvalidRequest(errors.New("date is required")))
-		return
-	}
-	date, err := berlinDate(req.Date)
-	if err != nil {
-		common.RenderError(w, r, common.ErrorInvalidRequest(errors.New("invalid date format, expected YYYY-MM-DD")))
-		return
-	}
-
-	// Past dates are rejected. /substitute is a planner tool for "today's staff
-	// is out" scenarios; mutating is_absent flags on completed instances would
-	// rewrite history. Matches the /gaps policy. Post-hoc correction (if ever
-	// required) belongs behind a distinct endpoint or an explicit flag, not
-	// this fast-path.
-	if date.Before(timezone.TodayDate()) {
-		common.RenderError(w, r, common.ErrorInvalidRequest(errors.New("'date' must be today or a future date")))
-		return
-	}
-
-	if rs.TimetableData == nil || rs.PersonService == nil {
-		common.RenderError(w, r, common.ErrorInternalServer(errors.New("timetable resource not fully wired")))
-		return
-	}
-
-	// Serialize the whole (tenant, date) before any classification read. A
-	// substitution/absence is day-wide (every same-day block of the absent staff),
-	// so a per-instance lock would let two admins picking different substitutes
-	// for the same employee both classify "no substitute" and each insert one,
-	// leaving two live supervisors on a shared block. The /deviations route takes
-	// the same day lock, so the two endpoints serialize against each other too.
-	// Covers both the substitute path below and the markAbsentOnly branch.
-	// Released automatically when this request's tenant tx commits/rolls back.
-	if err := rs.TimetableData.AcquireSubstituteDayLock(ctx, date); err != nil {
-		common.RenderError(w, r, common.ErrorInternalServerWrap("lock day failed", err))
-		return
-	}
-
-	// --- 404 checks: both staff must exist in this tenant -----------------
-	absentStaff, err := rs.PersonService.GetStaffByID(ctx, req.AbsentStaffID)
-	if err != nil {
-		if base.IsNoRows(err) {
-			common.RenderError(w, r, common.ErrorNotFound(errors.New("absent staff not found")))
-			return
-		}
-		common.RenderError(w, r, common.ErrorInternalServerWrap("load absent staff failed", err))
-		return
-	}
-	if absentStaff == nil || absentStaff.ID == 0 {
-		common.RenderError(w, r, common.ErrorNotFound(errors.New("absent staff not found")))
-		return
-	}
-
-	// Absent-only mode (#1840): no substitute given → mark the staff absent
-	// across the day and leave every affected position open. Handled by a
-	// dedicated helper so the classify/two-phase substitute path below stays
-	// exactly as-is.
-	if req.SubstituteStaffID == 0 {
-		rs.markAbsentOnly(w, r, req, date)
-		return
-	}
-
-	subStaff, err := rs.PersonService.GetStaffByID(ctx, req.SubstituteStaffID)
-	if err != nil {
-		if base.IsNoRows(err) {
-			common.RenderError(w, r, common.ErrorNotFound(errors.New("substitute staff not found")))
-			return
-		}
-		common.RenderError(w, r, common.ErrorInternalServerWrap("load substitute staff failed", err))
-		return
-	}
-	if subStaff == nil || subStaff.ID == 0 {
-		common.RenderError(w, r, common.ErrorNotFound(errors.New("substitute staff not found")))
-		return
-	}
-
-	// --- Reject a substitute who is themselves absent this day ------------
-	// Marking a staff member absent is day-wide (every same-day row flips
-	// is_absent=true), so a substitute with any absent row that day is out of
-	// action and cannot cover anyone — assigning them would create a non-absent
-	// substitute row that contradicts their own absent rows, and the overlap
-	// warning would never fire because absent rows are skipped. The frontend
-	// already hides day-absent staff from the picker; this is the backend guard.
-	subDayRows, err := rs.TimetableData.GetInstanceStaffByStaffAndDate(ctx, req.SubstituteStaffID, date)
-	if err != nil {
-		common.RenderError(w, r, common.ErrorInternalServerWrap("load substitute assignments failed", err))
-		return
-	}
-	for _, row := range subDayRows {
-		if row.IsAbsent {
-			common.RenderError(w, r, common.ErrorInvalidRequest(
-				errors.New("substitute is marked absent on this date")))
-			return
-		}
-	}
-
-	// --- Load absent staff's same-day assignments -------------------------
-	origRows, err := rs.TimetableData.GetInstanceStaffByStaffAndDate(ctx, req.AbsentStaffID, date)
-	if err != nil {
-		common.RenderError(w, r, common.ErrorInternalServerWrap("load absent assignments failed", err))
-		return
-	}
-
-	// No rows → empty success response. An absent staff without a scheduled
-	// day is not an error; responding 404 would conflate "staff exists?" with
-	// "staff has work today?".
-	if len(origRows) == 0 {
-		rs.getLogger().Info("substitute no-op (no assignments)",
-			slog.Int64("absent_staff_id", req.AbsentStaffID),
-			slog.Int64("substitute_staff_id", req.SubstituteStaffID),
-			slog.String("date", req.Date),
-		)
-		common.Respond(w, r, http.StatusOK, SubstituteResponse{
-			AbsentStaffID:     req.AbsentStaffID,
-			SubstituteStaffID: req.SubstituteStaffID,
-			Date:              req.Date,
-			AffectedInstances: []AffectedInstance{},
-			Warnings:          []scheduleSvc.SubstituteTimeConflict{},
-		}, "No assignments to substitute")
-		return
-	}
-
-	// ======================================================================
-	// PHASE A — Dry-Run: classify every row, no writes.
-	// ======================================================================
-	plan := make([]plannedOp, 0, len(origRows))
-
-	for _, orig := range origRows {
-		instance, err := rs.TimetableData.GetActivityInstance(ctx, orig.InstanceID)
-		if err != nil {
-			common.RenderError(w, r, common.ErrorInternalServerWrap("load target instance failed", err))
-			return
-		}
-		if instance == nil {
-			// Data-corruption branch: instance_staff row references an instance
-			// that no longer exists. Aborting is safer than silently skipping.
-			common.RenderError(w, r, common.ErrorInternalServer(
-				fmt.Errorf("instance_staff %d references missing instance %d", orig.ID, orig.InstanceID)))
-			return
-		}
-
-		// Skip terminal (completed/cancelled) instances: those are historical
-		// staffing that must not be rewritten. A same-day already-finished block
-		// is returned by the loader but is not a valid substitute target.
-		if !isPlannableInstance(instance) {
-			continue
-		}
-
-		// All rows of this instance — needed to detect existing substitutes
-		// and co-supervisor cases.
-		allRows, err := rs.TimetableData.GetInstanceStaff(ctx, instance.ID)
-		if err != nil {
-			common.RenderError(w, r, common.ErrorInternalServerWrap("load instance staff failed", err))
-			return
-		}
-
-		action, conflictOtherStaff, ok := classifySubstitute(allRows, orig, req.SubstituteStaffID)
-		if !ok {
-			// Phase A aborts the whole request — no writes have happened yet,
-			// so even though the tenant middleware will commit the (empty)
-			// tx, no DB state changed. The stable code lets the frontend
-			// render the conflict without parsing the German message.
-			common.RenderError(w, r, common.ErrorConflictWithCode(
-				fmt.Errorf("instance %d is already fully covered by another substitute (staff_id=%d); remove a replacement before adding one",
-					instance.ID, conflictOtherStaff),
-				"substitute_conflict",
-			))
-			return
-		}
-
-		plan = append(plan, plannedOp{
-			instance: instance,
-			origRow:  orig,
-			action:   action,
-		})
-	}
-
-	// ======================================================================
-	// PHASE B — Apply writes. Any DB error from here renders 500 → the
-	// middleware rolls the tx back. No 4xx path writes anything.
-	// ======================================================================
-	now := time.Now()
-	actor := resolveActorAccountID(ctx)
-	affected := make([]AffectedInstance, 0, len(plan))
-	// Collect active-group IDs we touched to fire SSE updates at the end.
-	activeTouched := make(map[int64]*scheduleModel.ActivityInstance)
-	// Instances that carried a stale "deliberately unstaffed" acknowledgement
-	// and now receive coverage — the ack must be cleared in the same tx, or the
-	// block reads amber-and-unstaffed while /gaps excludes it as staffed (#1840).
-	clearAck := make(map[int64]bool)
-	reason := trimReason(req.Reason)
-
-	for _, op := range plan {
-		if err := rs.InstanceService.ApplySubstitute(ctx, op.writeOp(), req.SubstituteStaffID, reason, now, actor, activeTouched); err != nil {
-			common.RenderError(w, r, common.ErrorInternalServerWrap("apply substitute failed", err))
-			return
-		}
-		// Coverage added (or confirmed) on a block still flagged unstaffed →
-		// mark it for acknowledgement clearing below.
-		if op.instance.UnderstaffedAck &&
-			(op.action == substituteActionSubstituted || op.action == substituteActionAlreadyOnInstance) {
-			clearAck[op.instance.ID] = true
-		}
-
-		affected = append(affected, AffectedInstance{
-			InstanceID: op.instance.ID,
-			Title:      op.instance.Title,
-			StartTime:  op.instance.StartTime.Format("15:04"),
-			Action:     op.action,
-		})
-	}
-
-	// Clearing the acknowledgement RECOMPUTES coverage first: a single
-	// replacement on a block with several open positions can still leave
-	// present < planned, so the service only clears the flag when the block is
-	// no longer understaffed. Otherwise partial coverage would silently reopen
-	// an intentionally acknowledged gap (#1840). Runs in this same tenant tx.
-	if rs.InstanceService != nil {
-		for instanceID := range clearAck {
-			if err := rs.InstanceService.ClearUnderstaffedAckIfStaffed(ctx, instanceID, actor); err != nil {
-				common.RenderError(w, r, common.ErrorInternalServerWrap("clear stale understaffed ack failed", err))
-				return
-			}
-		}
-	}
-
-	// ======================================================================
-	// Soft warnings: substitute's OTHER same-day assignments that overlap a
-	// target. Only relevant for instances actually substituted or co-cover.
-	// Dry-run: no DB writes, just a second Find + comparison.
-	// ======================================================================
-	warnings, err := rs.buildSubstituteTimeConflicts(ctx, plan, req.SubstituteStaffID, date)
-	if err != nil {
-		// A warning lookup failure is not a showstopper. Log and continue
-		// with an empty warning list — the mutations above are correct
-		// independent of the advisory output.
-		rs.getLogger().Warn("substitute time-conflict detection failed",
-			slog.Int64("substitute_staff_id", req.SubstituteStaffID),
-			slog.String("error", err.Error()),
-		)
-		warnings = nil
-	}
-	if warnings == nil {
-		warnings = []scheduleSvc.SubstituteTimeConflict{}
-	}
-
-	// ======================================================================
-	// SSE broadcast — fire-and-forget inside the tx, consistent with WP-B9.
-	// ======================================================================
-	rs.broadcastSubstituteEvents(ctx, activeTouched)
-
-	rs.getLogger().Info("substitute applied",
-		slog.Int64("absent_staff_id", req.AbsentStaffID),
-		slog.Int64("substitute_staff_id", req.SubstituteStaffID),
-		slog.String("date", req.Date),
-		slog.Int("affected_count", len(affected)),
-		slog.Int("warning_count", len(warnings)),
-	)
-
-	common.Respond(w, r, http.StatusOK, SubstituteResponse{
-		AbsentStaffID:     req.AbsentStaffID,
-		SubstituteStaffID: req.SubstituteStaffID,
-		Date:              req.Date,
-		AffectedInstances: affected,
-		Warnings:          warnings,
-	}, "Substitute applied")
-}
-
-// markAbsentOnly handles the absent-only branch of POST /substitute (#1840):
-// substitute_staff_id was omitted, so we mark the absent staff's same-day
-// assignments is_absent=true and leave the positions open. No substitute row is
-// created. For active instances the absent's live supervisor row is ended, same
-// as the substitute path. There is no 409 case here — marking absent is
-// idempotent and conflict-free — so a single write pass is safe (any error is a
-// 500 that the tenant middleware rolls back). Reuses the shared SSE broadcast.
-func (rs *Resource) markAbsentOnly(w http.ResponseWriter, r *http.Request, req substituteRequest, date timezone.Date) {
-	ctx := r.Context()
-
-	origRows, err := rs.TimetableData.GetInstanceStaffByStaffAndDate(ctx, req.AbsentStaffID, date)
-	if err != nil {
-		common.RenderError(w, r, common.ErrorInternalServerWrap("load absent assignments failed", err))
-		return
-	}
-	if len(origRows) == 0 {
-		rs.getLogger().Info("mark-absent no-op (no assignments)",
-			slog.Int64("absent_staff_id", req.AbsentStaffID),
-			slog.String("date", req.Date),
-		)
-		common.Respond(w, r, http.StatusOK, SubstituteResponse{
-			AbsentStaffID:     req.AbsentStaffID,
-			SubstituteStaffID: 0,
-			Date:              req.Date,
-			AffectedInstances: []AffectedInstance{},
-			Warnings:          []scheduleSvc.SubstituteTimeConflict{},
-		}, "No assignments to mark absent")
-		return
-	}
-
-	affected := make([]AffectedInstance, 0, len(origRows))
-	activeTouched := make(map[int64]*scheduleModel.ActivityInstance)
-	actor := resolveActorAccountID(ctx)
-
-	for _, orig := range origRows {
-		instance, err := rs.TimetableData.GetActivityInstance(ctx, orig.InstanceID)
-		if err != nil {
-			common.RenderError(w, r, common.ErrorInternalServerWrap("load target instance failed", err))
-			return
-		}
-		if instance == nil {
-			common.RenderError(w, r, common.ErrorInternalServer(
-				fmt.Errorf("instance_staff %d references missing instance %d", orig.ID, orig.InstanceID)))
-			return
-		}
-
-		// Skip terminal (completed/cancelled) instances so an already-finished
-		// same-day block is never retroactively marked absent. See
-		// isPlannableInstance.
-		if !isPlannableInstance(instance) {
-			continue
-		}
-
-		action := substituteActionMarkedAbsent
-		if orig.IsAbsent {
-			// Already absent — idempotent replay, no write, no protocol entry.
-			action = substituteActionAlreadyAbsent
-		} else if err := rs.InstanceService.ApplyAbsence(ctx, orig, instance, trimReason(req.Reason), actor, activeTouched); err != nil {
-			common.RenderError(w, r, common.ErrorInternalServerWrap("update original staff row failed", err))
-			return
-		}
-
-		affected = append(affected, AffectedInstance{
-			InstanceID: instance.ID,
-			Title:      instance.Title,
-			StartTime:  instance.StartTime.Format("15:04"),
-			Action:     action,
-		})
-	}
-
-	rs.broadcastSubstituteEvents(ctx, activeTouched)
-
-	rs.getLogger().Info("mark-absent applied",
-		slog.Int64("absent_staff_id", req.AbsentStaffID),
-		slog.String("date", req.Date),
-		slog.Int("affected_count", len(affected)),
-	)
-
-	common.Respond(w, r, http.StatusOK, SubstituteResponse{
-		AbsentStaffID:     req.AbsentStaffID,
-		SubstituteStaffID: 0,
-		Date:              req.Date,
-		AffectedInstances: affected,
-		Warnings:          []scheduleSvc.SubstituteTimeConflict{},
-	}, "Staff marked absent")
 }
 
 // trimReason normalizes an optional deviation reason: nil/blank becomes nil,
