--- frontend/src/app/api/timetable/substitute/route.ts
@@ -1,3 +1,0 @@
-import { proxyPost } from "~/lib/route-proxy.server";
-
-export const POST = proxyPost("/api/timetable/substitute");
