+++ frontend/src/components/timetable/plan-quality-panel.test.tsx
@@ -200,7 +200,7 @@ describe("PlanQualityPanel", () => {
     fireEvent.click(screen.getByRole("button", { name: "Anwenden" }));
 
     await waitFor(() =>
-      expect(onSubstitute).toHaveBeenCalledWith("11", "12", "2026-05-04"),
+      expect(onSubstitute).toHaveBeenCalledWith("42", "11", "12", "2026-05-04"),
     );
 
     fireEvent.click(screen.getAllByRole("button", { name: /Mensa/ })[1]!);
