+++ frontend/src/lib/timetable-api.ts
@@ -26,7 +26,6 @@ import type {
   BackendReplanWeekResult,
   BackendSplitTemplateResult,
   BackendStartInstanceResult,
-  BackendSubstituteResponse,
   BackendAcknowledgeUnderstaffedResponse,
   BackendApplyDeviationsResponse,
   BackendTimetableTemplate,
@@ -53,7 +52,6 @@ import type {
   SplitTemplateBody,
   SplitTemplateResult,
   StartInstanceResult,
-  SubstituteResponse,
   AcknowledgeUnderstaffedResponse,
   ApplyDeviationsInput,
   ApplyDeviationsResponse,
@@ -79,7 +77,6 @@ import {
   mapReplanWeekResult,
   mapSplitTemplateResult,
   mapStartInstanceResult,
-  mapSubstitute,
   mapTemplates,
   mapWeeklyInstances,
 } from "./timetable-helpers";
@@ -605,60 +602,6 @@ class TimetableService {
 
     const raw = await unwrap<BackendExceptionConflictsResponse>(response);
     return mapExceptionConflicts(raw);
-  }
-
-  async substitute(
-    absentStaffId: string,
-    substituteStaffId: string,
-    date: string,
-    reason?: string,
-  ): Promise<SubstituteResponse> {
-    const response = await fetch("/api/timetable/substitute", {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-        Accept: "application/json",
-      },
-      credentials: "include",
-      body: JSON.stringify({
-        absent_staff_id: Number(absentStaffId),
-        substitute_staff_id: Number(substituteStaffId),
-        date,
-        reason: reason ?? undefined,
-      }),
-    });
-
-    const raw = await unwrap<BackendSubstituteResponse>(response);
-    return mapSubstitute(raw);
-  }
-
-  /**
-   * Mark a staff member absent across their whole day without assigning a
-   * substitute (#1840 absent-only mode). The position is left open. Same
-   * endpoint as substitute — the backend branches on the omitted
-   * substitute_staff_id.
-   */
-  async markAbsent(
-    absentStaffId: string,
-    date: string,
-    reason?: string,
-  ): Promise<SubstituteResponse> {
-    const response = await fetch("/api/timetable/substitute", {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-        Accept: "application/json",
-      },
-      credentials: "include",
-      body: JSON.stringify({
-        absent_staff_id: Number(absentStaffId),
-        date,
-        reason: reason ?? undefined,
-      }),
-    });
-
-    const raw = await unwrap<BackendSubstituteResponse>(response);
-    return mapSubstitute(raw);
   }
 
   /**
