+++ frontend/src/lib/timetable-helpers.test.ts
@@ -39,7 +39,6 @@ import {
   mapReplanWeekResult,
   mapSplitTemplateResult,
   mapStartInstanceResult,
-  mapSubstitute,
   mapShiftCoverageCheckResult,
   mapTemplates,
   mapWeeklyInstances,
@@ -683,140 +682,6 @@ describe("backend mappers", () => {
       understaffedAck: true,
       understaffedNote: "keine Vertretung",
       cancelReason: "Ausflug",
-    });
-  });
-
-  it("maps substitutes and templates", () => {
-    expect(
-      mapSubstitute({
-        absent_staff_id: 11,
-        substitute_staff_id: 12,
-        date: "2026-05-04",
-        affected_instances: [
-          {
-            instance_id: 42,
-            title: "Mensa",
-            start_time: "12:00",
-            action: "substituted",
-          },
-        ],
-        warnings: [
-          {
-            instance_id: 43,
-            title: "Yoga",
-            date: "2026-05-04",
-            start_time: "14:00",
-            end_time: "15:00",
-          },
-        ],
-      }),
-    ).toMatchObject({
-      absentStaffId: "11",
-      substituteStaffId: "12",
-      affectedInstances: [{ instanceId: "42" }],
-      warnings: [{ instanceId: "43" }],
-    });
-
-    expect(
-      mapTemplates({
-        templates: [
-          {
-            id: 7,
-            name: "Yoga",
-            type: "activity",
-            category_id: 2,
-            category_name: "AG",
-            room_id: 3,
-            room_name: "Turnhalle",
-            is_open: true,
-            max_participants: 12,
-            target_group_type: "none",
-            enrollment_count: 8,
-            supervisor_count: 1,
-            required_staff_count: 1,
-            assigned_staff_count: 1,
-            student_ids: [21],
-            staff_ids: [11],
-            primary_staff_id: 11,
-            schedules: [
-              {
-                id: 9,
-                weekday: 1,
-                start_time: "14:00",
-                end_time: "15:00",
-                week_pattern: 2,
-                calendar_period_id: 5,
-                valid_until: "2026-06-01",
-              },
-            ],
-          },
-        ],
-      }).templates[0],
-    ).toMatchObject({
-      id: "7",
-      roomId: "3",
-      studentIds: ["21"],
-      staffIds: ["11"],
-      primaryStaffId: "11",
-      schedules: [{ id: "9", calendarPeriodId: "5", validUntil: "2026-06-01" }],
-    });
-  });
-
-  it("maps optional substitute/template fields without fallback ids", () => {
-    expect(
-      mapSubstitute({
-        absent_staff_id: 11,
-        substitute_staff_id: 12,
-        date: "2026-05-04",
-        affected_instances: undefined as never,
-        warnings: undefined as never,
-      }),
-    ).toMatchObject({ affectedInstances: [], warnings: [] });
-
-    expect(
-      mapTemplates({
-        templates: [
-          {
-            id: 8,
-            name: "Offene Betreuung",
-            type: "care",
-            category_id: 3,
-            category_name: "Betreuung",
-            is_open: false,
-            max_participants: undefined as never,
-            target_group_type: "none",
-            enrollment_count: undefined as never,
-            supervisor_count: 0,
-            required_staff_count: 0,
-            assigned_staff_count: 0,
-            schedules: undefined as never,
-          },
-        ],
-      }).templates[0],
-    ).toEqual({
-      id: "8",
-      name: "Offene Betreuung",
-      type: "care",
-      categoryId: "3",
-      categoryName: "Betreuung",
-      roomId: undefined,
-      roomName: undefined,
-      educationGroupId: undefined,
-      educationGroupName: undefined,
-      isOpen: false,
-      maxParticipants: undefined,
-      calendarPeriodId: undefined,
-      targetGroupType: "none",
-      targetGradeLevel: undefined,
-      targetSchoolClass: undefined,
-      enrollmentCount: undefined,
-      supervisorCount: 0,
-      requiredStaffCount: 0,
-      assignedStaffCount: 0,
-      studentIds: [],
-      staffIds: [],
-      primaryStaffId: undefined,
-      schedules: [],
     });
   });
 });
